+++ packages/types/src/index.ts
@@ -10,3 +10,4 @@ export * from './signal';
 export * from './discovery';
 export * from './artist-band';
 export * from './admin';
+export * from './print-shop';